Moved the main function away from ship.cpp into main.cpp file
